requirements.txt
Discord_Webhooks==1.0.4
Django==3.1.4
python-dotenv==0.15.0
django-crispy-forms==1.10.0asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
discord-webhooks==1.0.4
django-crispy-forms==1.11.2
django==3.1.4
djangorestframework==3.12.2
idna==2.8
python-dotenv==0.15.0
pytz==2021.1
requests==2.21.0
sqlparse==0.4.1
urllib3==1.24.3
asgiref==3.3.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
discord-webhooks==1.0.4
django-crispy-forms==1.11.2
django==3.1.4
djangorestframework==3.12.2
idna==2.8
pycparser==2.20
python-dotenv==0.15.0
pytz==2021.1
requests==2.21.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.24.3
